Verizon Premium CDN (Azure) not respecting token restrictions
The issue was that the CDN was not respecting the restrictions within the token. The reason was that the Verizon portal uses underscore "_" for the parameter names whereas this class was putting a dash "-". This is now fixed by aligning this "SDK" class with what is shown on the CDN Profile management screen in Azure.