Add bad_actor_tests + c# interop support
Adds tests to address #12

We didn't test the C# implementation along with all the other
implementations, but now we do.

There's some issue with the perl script when used from bad_actor_tests,
not sure what's up with that.

bad_actor_tests.py uses python_ectoken as a library, but the `-`
prevented us from importing it, so now it's moved to python_ectoken.